@@ -5,8 +5,8 @@ Gem::Specification.new do |s|
   s.name                      = "ae_page_objects"
   s.version                   = AePageObjects::VERSION
 
-  s.required_ruby_version     = '>= 1.9.3'
-  s.authors                   = ["Donnie Tognazzini"]
+  s.required_ruby_version     = '>= 2.2.5'
+  s.authors                   = ["AppFolio Engineering"]
   s.description               = "Capybara Page Objects pattern"
   s.email                     = ["engineering@appfolio.com"]
 
@@ -17,6 +17,6 @@ Gem::Specification.new do |s|
 
   s.files                     = `git ls-files -- lib`.split("\n")
 
-  s.add_dependency('capybara', '~> 2.1')
+  s.add_dependency('capybara', '~> 3.0')
 end
 
