@@ -24,7 +24,7 @@ module NodeInterfaceTests
     element = subject.element("#blahblah")
 
     capybara_node = stub(:allow_reload!)
-    capybara_stub.session.expects(:first).with("#blahblah").returns(capybara_node)
+    capybara_stub.session.expects(:first).with("#blahblah", minimum: 0).returns(capybara_node)
     verify_element(element, AePageObjects::Element, subject, capybara_node)
   end
 
@@ -34,7 +34,7 @@ module NodeInterfaceTests
     element = subject.element(:locator => ["#blahblah", {:visible => true}])
 
     capybara_node = stub(:allow_reload!)
-    capybara_stub.session.expects(:first).with("#blahblah", :visible => true).returns(capybara_node)
+    capybara_stub.session.expects(:first).with("#blahblah", visible: true, minimum: 0).returns(capybara_node)
     verify_element(element, AePageObjects::Element, subject, capybara_node)
   end
 
@@ -46,7 +46,7 @@ module NodeInterfaceTests
     element = subject.element(:locator => ["#blahblah", {:visible => true}], :is => element_class)
 
     capybara_node = stub(:allow_reload!)
-    capybara_stub.session.expects(:first).with("#blahblah", :visible => true).returns(capybara_node)
+    capybara_stub.session.expects(:first).with("#blahblah", visible: true, minimum: 0).returns(capybara_node)
     verify_element(element, element_class, subject, capybara_node)
   end
 
