@@ -265,6 +265,11 @@ class PageObjectIntegrationTest < Selenium::TestCase
   end
 
   def test_some_collection_enumerables
+    Author.delete_all
+    index = PageObjects::Authors::IndexPage.visit
+
+    assert_equal 0, index.authors.size
+
     ActiveRecord::Base.transaction do
       Author.create!(last_name: 'a')
       Author.create!(last_name: 'm')
@@ -276,7 +281,7 @@ class PageObjectIntegrationTest < Selenium::TestCase
 
     index = PageObjects::Authors::IndexPage.visit
 
-    assert_equal 8, index.authors.size
+    assert_equal 6, index.authors.size
     assert_nil index.authors.find { |author|
       author.last_name.text == 'q'
     }
