@@ -19,21 +19,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -55,22 +55,22 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
 
-        previous_owners_page_object.expects(:all).with(:xpath, ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath, ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owners_class.item_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -91,22 +91,22 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
 
-        previous_owners_page_object.expects(:all).with(:xpath, ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath, ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owners_class.item_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -130,21 +130,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -168,21 +168,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -204,21 +204,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -238,21 +238,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -274,21 +274,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -317,21 +317,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -345,13 +345,13 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#previous_owners").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
       end
 
@@ -370,21 +370,21 @@ module AePageObjects
         jon = kitty.new
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("whatever you want, baby").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("whatever you want, baby", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("Kitty Name").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("Kitty Name", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
 
@@ -405,21 +405,21 @@ module AePageObjects
         jon.expects(:page_local_context).returns("hello")
 
         previous_owners_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("hello").returns(previous_owners_page_object)
+        capybara_stub.session.expects(:first).with("hello", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
         first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
-        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", {}).returns(first_owner_page_object)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", wait: false).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
 
         owner_name_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name").returns(owner_name_page_object)
+        first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
         kitty_name_during_ownership_page_object = stub(:allow_reload!)
-        first_owner_page_object.expects(:first).with("Milkshake").returns(kitty_name_during_ownership_page_object)
+        first_owner_page_object.expects(:first).with("Milkshake", minimum: 0).returns(kitty_name_during_ownership_page_object)
 
         first_owner.expects(:page_local_context).returns("Milkshake")
 
