@@ -15,7 +15,7 @@ module AePageObjects
         jon = kitty.new
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#kind").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Element, kind_page_object)
       end
 
@@ -32,7 +32,7 @@ module AePageObjects
         jon = kitty.new
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("Kind Homie").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("Kind Homie", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Element, kind_page_object)
       end
 
@@ -50,7 +50,7 @@ module AePageObjects
         jon.expects(:page_local_context).returns("hello")
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("hello").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("hello", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Element, kind_page_object)
       end
 
@@ -67,7 +67,7 @@ module AePageObjects
         jon = kitty.new
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#kind").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Select, kind_page_object)
       end
 
@@ -84,7 +84,7 @@ module AePageObjects
         jon = kitty.new
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#kind").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Checkbox, kind_page_object)
       end
 
@@ -103,7 +103,7 @@ module AePageObjects
         jon = kitty.new
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#kind").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
 
         verify_element_on_parent(jon, :kind, special_widget, kind_page_object)
       end
@@ -123,7 +123,7 @@ module AePageObjects
         jon = kitty.new
 
         kind_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("As If!").returns(kind_page_object)
+        capybara_stub.session.expects(:first).with("As If!", minimum: 0).returns(kind_page_object)
 
         verify_element_on_parent(jon, :kind, special_widget, kind_page_object)
       end
@@ -150,26 +150,26 @@ module AePageObjects
         jon = kitty.new
 
         tail_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#tail_attributes").returns(tail_page_object)
+        capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, AePageObjects::Element, tail_page_object)
 
         color_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_color").returns(color_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
         size_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_size_attributes").returns(size_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
         length_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length").returns(length_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
         width_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_width").returns(width_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_width", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
 
@@ -197,22 +197,22 @@ module AePageObjects
         jon = kitty.new
 
         tail_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#tail_attributes").returns(tail_page_object)
+        capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent(jon, :tail, tail_class, tail_page_object)
 
         color_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_color").returns(color_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
         size_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_size_attributes").returns(size_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
         length_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length").returns(length_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
         width_page_object = stub(:allow_reload!)
@@ -244,26 +244,26 @@ module AePageObjects
         jon = kitty.new
 
         tail_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#tail_attributes").returns(tail_page_object)
+        capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, tail_base_class, tail_page_object)
 
         color_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_color").returns(color_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
         size_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_size_attributes").returns(size_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
         length_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length").returns(length_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
         width_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_width").returns(width_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_width", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
 
@@ -289,26 +289,26 @@ module AePageObjects
         jon = kitty.new
 
         tail_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("what ever you want, baby").returns(tail_page_object)
+        capybara_stub.session.expects(:first).with("what ever you want, baby", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, AePageObjects::Element, tail_page_object)
 
         color_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_color").returns(color_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
         size_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("Size").returns(size_page_object)
+        tail_page_object.expects(:first).with("Size", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
         length_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length").returns(length_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
         width_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("Fatness").returns(width_page_object)
+        size_page_object.expects(:first).with("Fatness", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
 
@@ -334,28 +334,28 @@ module AePageObjects
         jon = kitty.new
 
         tail_page_object = stub(:allow_reload!)
-        capybara_stub.session.expects(:first).with("#tail_attributes").returns(tail_page_object)
+        capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, AePageObjects::Element, tail_page_object)
 
         color_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_color").returns(color_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
         size_page_object = stub(:allow_reload!)
-        tail_page_object.expects(:first).with("#tail_attributes_size_attributes").returns(size_page_object)
+        tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
         length_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length").returns(length_page_object)
+        size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
         size.expects(:page_local_context).returns("hello")
 
         width_page_object = stub(:allow_reload!)
-        size_page_object.expects(:first).with("hello").returns(width_page_object)
+        size_page_object.expects(:first).with("hello", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
 
