@@ -19,7 +19,7 @@ module AePageObjects
 
         capybara_node = stub(:allow_reload!)
         capybara_stub
-        capybara_stub.session.stubs(:first).with('#foo').returns(capybara_node)
+        capybara_stub.session.stubs(:first).with('#foo', minimum: 0).returns(capybara_node)
 
         loaded.clear
         kitty_class.new(document_class.new, '#foo')
